backend/requirements.txt
# Flask framework
Flask>=2.0.0

# CORS handling for frontend communication
Flask-Cors>=3.0.0

# YAML parsing for configuration
PyYAML>=5.0

# OpenAPI parsing (add specific libraries later if needed, e.g., prance, openapi-spec-validator)
# requests>=2.0.0 # If fetching specs from URL
# GitPython>=3.0.0 # If fetching specs from Git repo directly

# Add other dependencies as needed
